Add option to upload (single) photo for pet in add_pet
Only backend is done. Plans are on to extend this so you can upload
many pics at once, but first let's see if at least single pic works.

*You can ignore the rest of the text, unless you really want to read*

There were a lot of things to figure out, mainly because this is
virtually the first time I'm handling file uploads on **any** web
framework, whatsoever. Except fro the admin interface, and that's
not really counted. And then, there's the whole thing about having
a ListField of ImageFields in MongoEngine. Turns out it's best to
have a proxy OneImage model, at least according to KoppeKTop at
https://github.com/MongoEngine/flask-mongoengine/issues/86. Thanks,
KoppeKTop!

Anyway, now on to multi-upload with Pupload. Let's see how that
goes.... And there's also the thing about *how* do I display an
image that's stored in GridFS once it gets in there? It'll all be
sorted out in good time, I guess.

Sorry for this long commit message: more like a diary-entry. It's
going to use up quite a few bytes. Luckily, we have lots to spare
...I *hope*